summary controller created and tested 1/9 11:29PM

+++ src/main/java/com/example/calorie_calendar/calendar/User.java
@@ -6,7 +6,7 @@ public class User {
     private final String name;
     private double weight;
     private int height;
-    private double bmr;
+    private int bmr;
     private int age;
     private WeeklyTotal week;
     private final Gender gender;
@@ -41,16 +41,16 @@ public class User {
         return week;
     }
 
-    public double getBmr() {
+    public int getBmr() {
         return bmr;
     }
    // For men: ((9.65 × (weight in lb/2.2)) + (573 × (height in inches * .0254)) – (5.08 × age in years) + 260) * 1.2
    //  For women: ((7.38 × (weight in lb/2.2)) + (607 × (height in inches * .0254)) – (2.31 × age in years) + 43) * 1.2
     public void calculateBMR(){
         if(gender == Gender.MALE){
-            this.bmr = ((9.65 * (weight/2.2)) + (573 * (height * .0254)) - (5.08 * age) + 260) * 1.2;
+            this.bmr = (int) (((9.65 * (weight/2.2)) + (573 * (height * .0254)) - (5.08 * age) + 260) * 1.2);
         } else if(gender == Gender.FEMALE){
-            this.bmr = ((7.38 * (weight/2.2)) + (607 * (height * .0254)) - (2.31 * age) + 43) * 1.2;
+            this.bmr = (int) (((7.38 * (weight/2.2)) + (607 * (height * .0254)) - (2.31 * age) + 43) * 1.2);
         }
     }
 }
